Season 3

+++ src/control/combat/encounter_manager.py
@@ -60,14 +60,14 @@ class EncounterManager(Service):
 
     BOSS_TYPE = {
         3: EnemyType.DADDY_P1,
-        # 6: None,
+        6: EnemyType.WEEB_BALL,
         # 9: None,
         # 12: None,
     }
     BOSS_LEVEL = {v: k for k, v in BOSS_TYPE.items()}
     BOSS_KEY = {
         3: ItemType.DADDY_KEY,
-        6: None,
+        6: ItemType.WEEB_KEY,
         9: None,
         12: None,
     }
@@ -139,6 +139,7 @@ class EncounterManager(Service):
                     )
                     if await self.context_needs_update_check(context):
                         return
+                    await self.refresh_round_overview(context)
                 if combat_event.combat_event_type not in [
                     CombatEventType.ENEMY_END_TURN,
                     CombatEventType.MEMBER_END_TURN,
@@ -792,14 +793,15 @@ class EncounterManager(Service):
                 if skill_status_effect.self_target:
                     status_effect_target = character
 
-                await self.status_effect_manager.apply_status(
-                    context,
-                    character,
-                    status_effect_target,
-                    skill_status_effect.status_effect_type,
-                    skill_status_effect.stacks,
-                    application_value,
-                )
+                if random.random() < skill_status_effect.application_chance:
+                    await self.status_effect_manager.apply_status(
+                        context,
+                        character,
+                        status_effect_target,
+                        skill_status_effect.status_effect_type,
+                        skill_status_effect.stacks,
+                        application_value,
+                    )
 
             event = CombatEvent(
                 datetime.datetime.now(),
@@ -1084,7 +1086,7 @@ class EncounterManager(Service):
         async for message in thread.history(limit=100):
             if len(message.embeds) == 1 and message.author.id == self.bot.user.id:
                 embed = message.embeds[0]
-                if embed.image.url is not None:
+                if embed.image.url is not None and embed.title is not None:
                     await message.delete()
                     break
 
@@ -1092,7 +1094,7 @@ class EncounterManager(Service):
         async for message in thread.history(limit=100):
             if len(message.embeds) == 1 and message.author.id == self.bot.user.id:
                 embed = message.embeds[0]
-                if embed.image.url is not None:
+                if embed.image.url is not None and embed.title is not None:
                     return message
         return None
 
@@ -1110,6 +1112,19 @@ class EncounterManager(Service):
         enemy_controller = self.controller.get_service(controller_class)
         await enemy_controller.intro(encounter_id)
         await self.refresh_encounter_thread(encounter_id)
+
+    async def refresh_round_overview(self, context: EncounterContext):
+        round_message = await self.context_loader.get_previous_turn_message(
+            context.thread
+        )
+        if round_message is not None:
+            round_embeds = round_message.embeds
+            cont = round_embeds[0].title == "Round Continued.."
+            round_embed = await self.embed_manager.get_round_embed(context, cont=cont)
+            round_embeds[0] = round_embed
+            await self.context_loader.edit_message(
+                round_message, embeds=round_embeds, attachments=[]
+            )
 
     async def refresh_encounter_thread(self, encounter_id: int):
         context = await self.context_loader.load_encounter_context(encounter_id)
@@ -1146,17 +1161,7 @@ class EncounterManager(Service):
             await self.controller.dispatch_event(event)
             return
 
-        round_message = await self.context_loader.get_previous_turn_message(
-            context.thread
-        )
-        if round_message is not None:
-            round_embeds = round_message.embeds
-            cont = round_embeds[0].title == "Round Continued.."
-            round_embed = await self.embed_manager.get_round_embed(context, cont=cont)
-            round_embeds[0] = round_embed
-            await self.context_loader.edit_message(
-                round_message, embeds=round_embeds, attachments=[]
-            )
+        await self.refresh_round_overview(context)
 
         if not context.new_turn():
             return
